アップロード画面のリファクタリング
